Move all search builders into the Hyrax namespace
For consistency
